backend/requirements.txt
roboflow
ultralytics
flask
numpy
requests
psycopg2-binary